Rename of Meeg.Kentico.ContentComponents projects
The dependency on Kentico.Libraries meant that all refs to the lib folder to assemblies in that package were replaced - not ideal. Meeg.Kentico.ContentComponents is now a metapackage and Meeg.Kentico.ContentComponents.Core is a dependency for the metapackage and for the CMS package.

There were two reasons for this - one was to remove the dependency on `Kentico.Libraries` from the "core" project so it didn't get installed into consumer's CMSApp projects. The other was to bring it inline with the (evolving) naming conventions used on the KenticoContrib project.